pyproject.toml
[project]
name = "image-generation-papers"
version = "0.1.0"
description = "Implementations of image generation papers"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "einops==0.8.1",
    "hydra-core==1.3.2",
    "ipykernel==7.0.1",
    "ipywidgets==8.1.7",
    "numpy==2.3.4",
    "pandas==2.3.3",
    "pre-commit==4.3.0",
    "pydantic==2.11.9",
    "ruff==0.14.0",
    "safetensors==0.6.2",
    "torch==2.9.0",
    "torchmetrics[image]==1.8.2",
    "torchvision==0.24.0",
    "tqdm==4.67.1",
    "transformers==4.57.1",
    "wandb==0.22.2",
]

[tool.uv]
package = true

[tool.setuptools]
packages = ["src"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff]
line-length = 100
